+++ src/libsodium/crypto_xof/shake256/ref/shake256_ref.h
@@ -15,16 +15,13 @@ typedef struct shake256_state_internal_ {
     unsigned char domain; /* Domain separation byte */
 } shake256_state_internal;
 
-int shake256_ref(unsigned char *out, size_t outlen, const unsigned char *in,
-                 unsigned long long inlen);
+int shake256_ref(unsigned char *out, size_t outlen, const unsigned char *in, size_t inlen);
 
 int shake256_ref_init(shake256_state_internal *state);
 
 int shake256_ref_init_with_domain(shake256_state_internal *state, unsigned char domain);
 
-int shake256_ref_update(shake256_state_internal *state,
-                        const unsigned char     *in,
-                        unsigned long long       inlen);
+int shake256_ref_update(shake256_state_internal *state, const unsigned char *in, size_t inlen);
 
 int shake256_ref_squeeze(shake256_state_internal *state, unsigned char *out, size_t outlen);
